fixup! refactor: contract_master
