+++ contract_master/common/contract_master.py
@@ -5,7 +5,6 @@ from typing import Callable
 from pydantic.main import BaseModel
 
 from .models import CovalentTx, ServiceItem
-from .utils import equals, lower, unique
 
 
 class BalanceResult(BaseModel):
@@ -42,11 +41,7 @@ class ContractMaster(ABC):
         self.user_address = user_address
         self.txs = list(filter(self.__is_transaction_within(target_datetime), txs))
         self.block_height = self.__get_max_block_height(self.txs)
-        fungible_token_addresses, possessable_addresses = self.__get_relevant_contract_addresses(
-            base_address=user_address, transactions=self.txs
-        )
-        self.fungible_token_addresses = fungible_token_addresses
-        self.possessable_addresses = possessable_addresses
+
         pass
 
     def __is_transaction_within(self, max_datetime: datetime) -> Callable[[CovalentTx], bool]:
@@ -64,23 +59,3 @@ class ContractMaster(ABC):
     @abstractmethod
     def get_balances(self) -> list[list[BalanceResult] | list[IgnoredResult] | list[ErroredResult]]:
         pass
-
-    def __get_relevant_contract_addresses(
-        self, base_address: str, transactions: list[CovalentTx]
-    ) -> tuple[list[str], list[str]]:
-        fungible_token_addresses: list[str] = []
-        possessable_addresses: list[str] = []
-
-        for tx in transactions:
-            for e in tx.log_events:
-                if e.decoded and e.decoded.name == "Transfer":
-                    # 自分が含まれるTransferイベントのsender_addressはfungible tokenとして全て収集し、資産取得対象に含める
-                    if equals(e.decoded.get_param("from"), base_address) or equals(
-                        e.decoded.get_param("to"), base_address
-                    ):
-                        fungible_token_addresses.append(e.sender_address)
-                    # 自分からどこかのコントラクトにTransferしているものはpossessableとみなし、資産取得対象に含める
-                    if equals(e.decoded.get_param("from"), base_address):
-                        possessable_addresses.append(e.decoded.get_param("to"))
-
-        return unique(lower(fungible_token_addresses)), unique(lower(possessable_addresses))
